requirements.txt
pytorch
torchvision
matplotlib
tqdm
